Make proxy-deps multi-stage to remove the original source files
Previously, proxy-deps and go-deps included the source tree for local
projects. This can cause build conflicts when files are renamed.

By adopting a multi-stage build for the proxy-deps image, we can be sure
that we only preserve essential dependencies & manifests in the
proxy-deps and go-deps images.

Furthermore, bin/update-go-deps-shas and bin/update-proxy-deps-shas have
been added to ease maintenance when files are changed.

Fix #159

Signed-off-by: Oliver Gould <ver@buoyant.io>