Destinations: Recognize ".svc.cluster.local" and ".svc"
Previously these alternative ways of addressing a service were not
recognized by Conduit. Kubernetes doesn't guarantee these ways to
work but in practice some people are relying on them working since
minikube and some (all?) cloud providers, at least, support them.

Recognize these names in the Destinations service.

Later, we'll likely change the mechanics of this, especially when
TLS support is added, but this will work for now and that future change
should be transparent. The way this was done makes it easy to do that
behavior since it doesn't touch any of the configuration files for the
controller (`conduit install`) or for the proxies (`conduit inject`).

The existing unit tests were updated and expanded for the new
behavior.

Signed-off-by: Brian Smith <brian@briansmith.org>