Test the proxy in release mode in Docker in CI on the master branch.
Previously we were not running the proxy tests in the release configuration.

Run the proxy tests in the release configuration through Docker.

Docker builds with tests in release mode are too slow to run on every
pull request so release mode tests will only be run on the master
branch.

Signed-off-by: Brian Smith <brian@briansmith.org>