Improve the caching behavior of web/Dockerfile.
Previously if anything changed under web/ then the images would get
completely rebuilt. In particular fetching new packages in
`yarn install` can be very slow.

Attempt to minimize file dependencies in earlier stages of the Dockerfile
to maximize the number of layers that can be cached, both in the web
asset stage and the go server stage.

Signed-off-by: Brian Smith <brian@briansmith.org>