[rfc] Describe the Life of a Connection
We've been running into some ambiguities with regard to transparency &
service discovery.

This document _aspirationally_ describes the behavior of the proxy
(and explicitly not its current behavior).

My hope is that we can pick apart this larger narrative, agree on the
right end state, and turn that into a technical plan.

I don't expect this doc to merge in its current form.