Remove report-checking from ListPods
Currently, the ListPods implementation uses prometheus to see if the
instance has sent a telemetry report to the controller.

In preparation of removing the telemetry push functionality, the
`ListPods` endpoint should be changed to not rely on push telemetry.

While doing this, it seems that we don't actually need to talk to
prometheus to determine whether a pod is `added`. Instead, we check that
the pod is configured with the correct controller namespace.