proxy: Impose a hard upper limit on router capacity
Currently, the proxy may cache an unbounded number of routes. In order
to precent such leaks in production, new configurations are introduced
to limit the number of inbound and outbound HTTP routes. By default, we
support 100 inbound routes and 1000 outbound routes.

In a followup, we'll introduce an eviction strategy so that capacity can
be reclaimed gracefully.